Move AdminBundleIT to a separate profile
The IT was turned into an FT that will run
only in the FT profile. For now this is acceptable.
